move Position to constants and update imports
