+++ src/autocomplete/src/Autocomplete.js
@@ -4,7 +4,7 @@ import fuzzaldrin from 'fuzzaldrin-plus'
 import Downshift from 'downshift'
 import VirtualList from 'react-tiny-virtual-list'
 import { Popover } from '../../popover'
-import { Position } from '../../positioner'
+import { Position } from '../../constants'
 import { Heading } from '../../typography'
 import { Pane } from '../../layers'
 import AutocompleteItem from './AutocompleteItem'
