+++ src/constants/index.js
@@ -1,2 +1,3 @@
 export { default as StackingOrder } from './src/StackingOrder'
 export { default as Intent } from './src/Intent'
+export { default as Position } from './src/Position'
