+++ src/menu/stories/index.stories.js
@@ -6,7 +6,7 @@ import Menu from '../src/Menu'
 import { Button } from '../../buttons'
 import { UnorderedList, ListItem } from '../../typography'
 import { Popover } from '../../popover'
-import { Position } from '../../positioner'
+import { Position } from '../../constants'
 
 storiesOf('menu', module)
   .add('dropdown', () => (
