+++ src/popover/stories/index.stories.js
@@ -8,7 +8,7 @@ import { TextInputField } from '../../text-input'
 import { Pane } from '../../layers'
 import { Heading, Paragraph, Text } from '../../typography'
 import { Button } from '../../buttons'
-import { Position } from '../../positioner'
+import { Position } from '../../constants'
 import { Icon, IconNames } from '../../icon'
 
 const PopoverContent = () => (
