+++ src/table/stories/AdvancedTable.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { filter } from 'fuzzaldrin-plus'
 import { Table } from '../../table'
 import { Popover } from '../../popover'
-import { Position } from '../../positioner'
+import { Position } from '../../constants'
 import { Menu } from '../../menu'
 import { Avatar } from '../../avatar'
 import { Text } from '../../typography'
